add at least a simple does-it-run test for log_infeasible_constraints
